refactor the payment module from REST endpoint to graphQL call

+++ payment/src/main/kotlin/nl/nlportal/payment/domain/PaymentRequest.kt
@@ -15,10 +15,8 @@
  */
 package nl.nlportal.payment.domain
 
-import java.math.BigDecimal
-
 data class PaymentRequest(
-    val amount: BigDecimal,
+    val amount: Double,
     val reference: String,
     val orderId: String,
     val title: String?,
